// dashboard/src/utils/hasura-api/generated/schemas/createScheduledEventArgs.ts
/**
 * Generated by orval v8.4.0 🍺
 * Do not edit manually.
 * Hasura API
 * API for managing remote schemas and events in Hasura
 * OpenAPI spec version: 1.0.0
 */
import type { CreateScheduledEventArgsPayload } from './createScheduledEventArgsPayload';
import type { Headers } from './headers';
import type { RetryConfCT } from './retryConfCT';

export interface CreateScheduledEventArgs {
  /** URL of the webhook. A String value which supports templating environment variables enclosed in {{ and }}. */
  webhook: string;
  /** The time at which the invocation should be invoked (ISO8601 format). */
  schedule_at: string;
  /** Any JSON payload which will be sent when the webhook is invoked. */
  payload?: CreateScheduledEventArgsPayload;
  /** List of headers to be sent with the webhook */
  headers?: Headers;
  /** Retry configuration if scheduled event delivery fails */
  retry_conf?: RetryConfCT;
  /** Custom comment. */
  comment?: string;
}