// dashboard/src/utils/hasura-api/generated/schemas/retryConfCT.ts
/**
 * Generated by orval v7.18.0 🍺
 * Do not edit manually.
 * Hasura API
 * API for managing remote schemas and events in Hasura
 * OpenAPI spec version: 1.0.0
 */

/**
 * Retry configuration when delivery to the webhook of a cron trigger fails
 */
export interface RetryConfCT {
  /** Number of times to retry delivery. */
  num_retries?: number;
  /** Number of seconds to wait between each retry. */
  retry_interval_seconds?: number;
  /** Number of seconds to wait for the webhook response before timing out. */
  timeout_seconds?: number;
  /** Number of seconds between scheduled time and actual delivery time that is acceptable. If the time difference is more than this, then the event is dropped. */
  tolerance_seconds?: number;
}